+++ README.md
@@ -8,16 +8,16 @@
 
 
 ## 로또 기능 요구사항
-- [ ] 구입 금액을 입력받을 수 있다.
+- [x] 구입 금액을 입력받을 수 있다.
   - [x] 구입 금액은 1000 단위로만 입력한다.
-- [ ] 구입한 로또 갯수를 출력한다.
-- [ ] 구입한 로또 갯수 만큼 로또 번호 6개를 발급 후 출력한다.
+- [x] 구입한 로또 갯수를 출력한다.
+- [x] 구입한 로또 갯수 만큼 로또 번호 6개를 발급 후 출력한다.
   - [x] 로또 번호는 1부터 45까지이다.
   - [x] 한 로또에 중복되는 번호는 없다.
-  - [ ] 한 로또의 번호 출력은 오름차순이다.
-- [ ] 당첨 번호 6개와 보너스 번호 1개를 입력받을 수 있다.
+  - [x] 한 로또의 번호 출력은 오름차순이다.
+- [x] 당첨 번호 6개와 보너스 번호 1개를 입력받을 수 있다.
   - [x] 당첨 번호와 보너스 번호는 반드시 숫자여야 한다.
-  - [ ] 당첨 번호와 보너스 번호는 빈 문자열이거나 null 일 수 없다.
+  - [x] 당첨 번호와 보너스 번호는 빈 문자열이거나 null 일 수 없다.
   - [x] 로또 번호는 1부터 45까지이다.
   - [x] 당첨번호와 보너스 번호 모두 중복되는 번호는 없다.
   - [x] 당첨번호가 6개가 아니면 예외를 던진다.
@@ -30,23 +30,30 @@
   - [x] 4등 : 숫자 4개
   - [x] 5등 : 숫자 3개
   - [x] 꽝 : 숫자 2개 ~ 숫자 0개
-- [ ] 전체 발급된 로또의 통계를 출력한다.
+- [x] 전체 발급된 로또의 통계를 출력한다.
   - [x] 모든 당첨 로또의 상금 합 / 구입 금액 이 수익률이다.
+- [x] 예외 발생시 재입력 여부를 묻고 재입력할 경우 다시 입력받는다.
+  - [x] 재입력 여부는 Y 혹은 y를 입력시 재입력 의사가 있다고 판단한다.
+  - [x] N 혹은 n 입력시 재입력 의사 없다고 판단하고 종료한다.
+  - [x] 그 외 입력은 다시 재입력 여부를 묻는다.
 
 
 ## 객체 키워드
 - 구입 금액
 - 로또
-- 로또 번호
+- 개별 로또 번호
 - 당첨 번호 + 보너스 번호
 - 입력 parser -> `, `로 구분
 - 랭크(열거형)
 - 통계 내는 역할하는 애
-- 통계 데이터를 전달하는 dto
+- 통계 데이터를 전달하는 dto (미구현)
 - 로또 기계 -> 전체적인 흐름제어
 
 
 ## 시간되면 한번 고민
-- model.Money.multiply에 매개변수 타입이 Integer 말고 Wrapper Class로 변경
-- 로또 번호를 감쌀것인가? vs winningNumbers를 사용해서 보너스넘버도 같이 감쌀것인가?
+- parser가 뷰에 속한다고 판단하여 배치했는데, 도매인 객체 생성도 하고 있다. 이게 맞는건가?
+- 컨트롤러 역할을 Application이 하고 있다. 따로 여러개의 컨트롤러를 분리해야 될까?
+- Parser는 템플릿 메서드 패턴을 통해 보너스 번호, 로또 번호, 입금액을 파싱하고 있다. 코드 재활용은 좋아졌는데 각 클래스를 읽으면 이해가 되지 않는다.
+- 예외 재입력 과정에서 지나치게 많은 매개변수를 활용하고 있는 것 아닌가 싶다.
+- 예외가 에러 메시지, 즉 사용자에게 보여질 메시지를 관리하는 건 뷰에 관여하는 것 아닌가라는 생각이 들어서 따로 분리해서 하고자 했다.
