+++ src/main/java/model/Money.java
@@ -44,11 +44,4 @@ public class Money {
     public int hashCode() {
         return Objects.hash(amount);
     }
-
-    @Override
-    public String toString() {
-        return "model.Money{" +
-                "amount=" + amount +
-                '}';
-    }
 }
