+++ src/main/java/model/generator/LottoNumbersGenerator.java
@@ -1,4 +1,6 @@
-package model;
+package model.generator;
+
+import model.LottoNumbers;
 
 public interface LottoNumbersGenerator {
     LottoNumbers createLottoNumbers();
