+++ src/main/java/view/parser/MoneyParser.java
@@ -1,4 +1,4 @@
-package view;
+package view.parser;
 
 import model.Money;
