+++ src/test/java/view/parser/BonusNumberParserTest.java
@@ -1,13 +1,12 @@
-package view;
+package view.parser;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static view.BonusNumberParser.INVALID_BONUS_NUMBER_FORMAT_MESSAGE;
+import static view.parser.BonusNumberParser.INVALID_BONUS_NUMBER_FORMAT_MESSAGE;
 
 import exception.InvalidRangeLottoNumberException;
 import model.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
